fixed error in code, slight performance improvement
j was set to increment by 4 every loop, instead of 2. this means that it would have counted some numbers as prime when it shouldn't.

dividing the number of checks per number by 2 should increase performance a bit, as well as not impact accuracy as all numbers tested are odd anyways.